Bug fix. Refactor.
