+++ monoruby/src/tests.rs
@@ -436,11 +436,11 @@ mod test {
         ];
         let rhs = [
             "17",
-            "3454",
+            "5375",
             "25084",
             "234234645",
             "2352354645657876868978696835652452546462456245646",
-            "169.5333",
+            "234.2345",
             "690426.0",
         ];
         run_binop_tests(
